Use network=host on docker buildx build
We are already creating a custom buildx builder with options set to use
host networking. However, there are reports that this isn't sufficient
to get builds working in all situations. It is reported that adding
--network=host to the build command itself does fix things. Considering
we already intend to use host networking (per the custom buildx builder
options) lets go ahead and set --network=host on build commands as well.

This upstream issue may be related:
  https://github.com/docker/buildx/issues/1688
to the problems that have been observed, but it doesn't seem to match
completely. Maybe upstream behaviors changed since this was origianlly
reported?

Change-Id: Ie8856cf49b86e89e2bc903033e7a20874b7c9f83